requirements.txt
# Core dependencies
pandas==2.3.0              # pandas-dev
numpy==1.26.4              # numpy team
scikit-learn==1.7.0        # scikit-learn developers
seaborn==0.13.2            # Michael Waskom and contributors
wandb>=0.20.1              # Weights & Biases, Inc. (new API keys supported in newer versions)
# MetaTrader5==5.0.5120  isnt supported by WSL     # MetaQuotes Software Corp. (Windows only, Python 3.7–3.11)
python-dotenv==1.1.0       # Saurabh Kumar

# Deep learning dependencies (CUDA for Nvidia GPUs, e.g., RTX 3070)
torch==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121  # PyTorch with CUDA 12.1
torchvision==0.20.1+cu121 --index-url https://download.pytorch.org/whl/cu121  # For image processing
torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121  # Optional

# Optional/other dependencies
mypy==1.16.1               # Jukka Lehtosalo and contributors

# Machine Learning & Deep Learning
stable-baselines3==2.1.0
gymnasium==0.28.1          # Farama Foundation
shimmy==0.2.1
pytorch-lightning==2.0.2
optuna==3.2.0

# Monitoring & Visualization
plotly==5.14.1

# GPU Acceleration
pyopencl==2023.1.4
mako==1.2.4
pytools==2023.1.1

# Utilities
tqdm==4.65.0
requests==2.31.0
joblib==1.3.1
python-dateutil==2.8.2
pytz==2023.3
lxml==4.9.3

# Data API
dukascopy-python==4.0.1    # Dukascopy data downloader
